requirements.txt
psycopg-binary
psycopg
asyncpg
SQLAlchemy==2.0.30
alembic==1.13.1
boto3==1.34.128
python-dotenv==1.0.1
cryptography
PyMySQL==1.1.1
openpyxl==3.1.4

